flatpak: Upgrade from 1.12.7 to 1.13.3 (pre-release)
The 1.13 pre-release series has a lot of promising changes that
have made Flatpak a lot easier to package. It supports curl as
an HTTP backend instead of libsoup, so that's one package we can
remove. It incorporates a fix for the following bug so we don't
have to package xdg-dbus-proxy separately:
https://github.com/flatpak/xdg-dbus-proxy/issues/36

It also replaces appstream-glib with appstream.

This upgrade contains a custom patch to remove HTTP2 support because
KISS curl is built without it. In ostree this is done in the build
configuration but Flatpak doesn't support that.